+++ tests/requirements.py
@@ -31,6 +31,7 @@ class Requirements:
     DATUM_BUG_257 = "Dataset.filter doesn't count removed items"
     DATUM_BUG_259 = "Dataset.filter fails on merged datasets"
     DATUM_BUG_314 = "Unsuccessful remap_labels"
+    DATUM_BUG_402 = "Troubles running 'remap_labels' on ProjectDataset"
     DATUM_BUG_404 = "custom importer/extractor not loading"